+++ src/main/java/io/dataspaceconnector/service/message/handler/processor/DataRequestProcessor.java
@@ -27,6 +27,8 @@ import io.dataspaceconnector.service.message.builder.type.ArtifactResponseServic
 import io.dataspaceconnector.service.message.handler.dto.Response;
 import io.dataspaceconnector.service.message.handler.dto.RouteMsg;
 import io.dataspaceconnector.service.message.handler.processor.base.IdsProcessor;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -64,7 +66,7 @@ class DataRequestProcessor extends IdsProcessor<
      */
     @Override
     protected Response processInternal(final RouteMsg<ArtifactRequestMessageImpl,
-            MessagePayload> msg) throws Exception {
+            MessagePayload> msg, final Jws<Claims> claims) throws Exception {
         final var artifact = MessageUtils.extractRequestedArtifact(msg.getHeader());
         final var issuer = MessageUtils.extractIssuerConnector(msg.getHeader());
         final var messageId = MessageUtils.extractMessageId(msg.getHeader());