+++ src/main/java/io/dataspaceconnector/service/message/handler/processor/RejectContractProcessor.java
@@ -24,6 +24,8 @@ import io.dataspaceconnector.service.message.handler.dto.Response;
 import io.dataspaceconnector.service.message.handler.dto.RouteMsg;
 import io.dataspaceconnector.service.message.handler.dto.payload.ContractTargetRuleMapContainer;
 import io.dataspaceconnector.service.message.handler.processor.base.IdsProcessor;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -50,7 +52,7 @@ class RejectContractProcessor extends
      */
     @Override
     protected Response processInternal(final RouteMsg<ContractRequestMessageImpl,
-            ContractTargetRuleMapContainer> msg) throws Exception {
+            ContractTargetRuleMapContainer> msg, final Jws<Claims> claims) throws Exception {
         final var issuer = MessageUtils.extractIssuerConnector(msg.getHeader());
         final var messageId = MessageUtils.extractMessageId(msg.getHeader());