+++ src/test/java/io/dataspaceconnector/service/message/handler/event/AgreementEventHandlerTest.java
@@ -17,8 +17,12 @@ package io.dataspaceconnector.service.message.handler.event;
 
 import de.fraunhofer.iais.eis.ContractAgreementBuilder;
 import de.fraunhofer.ids.messaging.util.IdsMessageUtils;
+import io.dataspaceconnector.service.message.handler.dto.payload.AgreementClaimsContainer;
 import io.dataspaceconnector.service.usagecontrol.PolicyExecutionService;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -28,6 +32,9 @@ import static org.mockito.ArgumentMatchers.eq;
 
 @SpringBootTest(classes = {AgreementEventHandler.class})
 class AgreementEventHandlerTest {
+
+    @Mock
+    private Jws<Claims> claims;
 
     @MockBean
     private PolicyExecutionService executionService;
@@ -42,10 +49,13 @@ class AgreementEventHandlerTest {
                 ._contractStart_(IdsMessageUtils.getGregorianNow())
                 .build();
 
+        final var agreementContainer = new AgreementClaimsContainer(agreement, claims);
+
         /* ACT */
-        eventHandler.handleAgreementEvent(agreement);
+        eventHandler.handleAgreementEvent(agreementContainer);
 
         /* ASSERT */
-        Mockito.verify(executionService, Mockito.atLeastOnce()).sendAgreement(eq(agreement));
+        Mockito.verify(executionService, Mockito.atLeastOnce())
+                .sendAgreement(eq(agreement), eq(claims));
     }
 }